The import was failing because this needs to be an Integer
Error in question:
django.contrib.gis.utils.layermapping.LayerMapError: OGR field "<class 'django.contrib.gis.gdal.field.OFTString'>" (of type OFTString) cannot be mapped to Django FloatField.

Example of the mapping allowed values (from: http://phairs-devel.unm.edu/DREAM_DAILY_STAT_SERVER/epht_enviro_virtualenv/lib/python2.6/site-packages/django/contrib/gis/utils/layermapping.py):
        models.IntegerField : (OFTInteger, OFTReal, OFTString),

+++ django_zillow_neighborhoods/models.py
@@ -7,7 +7,7 @@ class Neighborhood(models.Model):
     county = models.CharField(max_length=43)
     city = models.CharField(max_length=64)
     name = models.CharField(max_length=64)
-    regionid = models.FloatField()
+    regionid = models.IntegerField()
     geom = models.MultiPolygonField(geography=True)
 
     objects = GeoManager()
